refactor(knockdog): api 레이어 model명 명확하게 변경
