+++ apps/knockdog/src/entities/kindergarten/api/kindergarten-main.ts
@@ -1,15 +1,15 @@
 import { api, type ApiResponse } from '@shared/api';
 
-import type { Kindergarten } from '../model/kindergarten';
+import type { Kindergarten } from '../model/search-list';
 
 // @TODO API Response, 타입 정의 필요
-export interface KindergartenMainRequest {
+export interface KindergartenMainParams {
   id: string;
   lng: number;
   lat: number;
 }
 
-function getKindergartenMain({ id, lng, lat }: KindergartenMainRequest): Promise<Kindergarten> {
+function getKindergartenMain({ id, lng, lat }: KindergartenMainParams): Promise<Kindergarten> {
   return api
     .get(`kindergarten/main/${id}`, {
       searchParams: {
