+++ apps/knockdog/src/entities/kindergarten/index.ts
@@ -3,8 +3,10 @@ export { getKindergartenAutocomplete } from './api/autocomplete';
 export type { KindergartenAutocompleteParams } from './api/autocomplete';
 export { getFilterResultCount } from './api/filters';
 export type { FilterResultCountParams } from './api/filters';
-export { getKindergartenSearchList, getKindergartenAggregation } from './api/kindergarten';
-export type { KindergartenSearchListParams, KindergartenAggregationParams } from './api/kindergarten';
+export { getKindergartenSearchList, getKindergartenAggregation } from './api/search-list';
+export type { KindergartenSearchListParams, KindergartenAggregationParams, SortType } from './api/search-list';
+export { getKindergartenMain } from './api/kindergarten-main';
+export type { KindergartenMainParams } from './api/kindergarten-main';
 
 /** config */
 export { ServiceTag, PickupType, OperationStatus } from './config/constant';
@@ -14,11 +16,14 @@ export { createKindergartenBasicQueryOptions } from './config/kindergartenBasicQ
 export { createKindergartenNearQueryOptions, kindergartenNearQueryKeys } from './config/kindergartenNearQueryKeys';
 export { createKindergartenMainQueryOptions, kindergartenMainQueryKeys } from './config/kindergartenMainQueryKeys';
 
+/** lib */
+export { isValidLatLngBounds, isSameBounds } from './lib/is';
+
 /** model */
 export { createKindergartenListWithMock } from './model/mappers';
-export type { Autocomplete, RegionSuggestion, FilterItemSuggestion, FilterItemCode } from './model/autocomplete';
+export type { Autocomplete, RegionSuggestion, FilterItemSuggestion, FilterItemCode, Place } from './model/autocomplete';
 export type { FilterResultCount } from './model/filters';
-export type { Aggregation, Kindergarten, SidoGunguAggregation } from './model/kindergarten';
+export type { Aggregation, Kindergarten, SidoGunguAggregation } from './model/search-list';
 export type { KindergartenNear } from './model/kindergarten-near';
 export type { KindergartenBasic, Coord, OperationTime } from './model/kindergarten-basic';
 export type {
